no message
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@96338 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0aa23858ae9d6394b450886129411c6dde594afd

+++ Online/Online/CCPCHist/src/dimhist.h
@@ -5,24 +5,24 @@
 class DimHistbuff1
 {
 public:
-	int dim;
-	int nxbin;
+	float dim;
+	float nxbin;
 	float xmin;
 	float xmax;
-	int nentries;
+	float nentries;
 	bintype entries;
 };
 class DimHistbuff2 
 {
 public:
-	int dim;
-	int nxbin;
+	float dim;
+	float nxbin;
 	float xmin;
 	float xmax;
-	int nybin;
+	float nybin;
 	float ymin;
 	float ymax;
-	int nentries;
+	float nentries;
 	bintype entries;
 };
 class Histo;
